refactor: 출석 편집 권한 로직을 model로 분리 및 Leader 범위 명확화
- canEditUserAttendance, isInSubtree를 server/src/model/attendance.ts로 이동 (라우터 슬림화)
- Leader는 자기 community만, VillageLeader는 subtree 전체로 권한 범위를 명시적으로 분리 (기존엔 둘 다 subtree 체크)
- isInSubtree에 id→node Map을 추가해 탐색을 O(1)로 개선
- memo 미전송 시 기존 memo가 빈 문자열로 덮어써지던 문제를 조건부 업데이트로 수정
